+++ src/actions/auth-actions.ts
@@ -85,3 +85,44 @@ export async function VerifyEmailAction(tokenUUID: string) {
   }
 }
 
+
+
+export async function ForgotPasswordAction(email: string) {
+  try {
+    const response = await apiClient.post("/api/v2.0/auth/password-reset/", {
+      email,
+    });
+    return { success: true, detail: response.data.detail };
+  } catch (error: any) {
+    if (error.response?.data?.detail) {
+      return { error: error.response.data.detail };
+    }
+
+    // Handle specific field errors if any (though usually it's just detail for this endpoint)
+    if (error.response?.data?.email) {
+      return { error: error.response.data.email[0] }
+    }
+
+    return { error: "Ocurrió un error al procesar tu solicitud." };
+  }
+}
+
+export async function ResetPasswordAction(uid: string, token: string, newPassword: string) {
+  try {
+    const response = await apiClient.post("/api/v2.0/auth/password-reset-confirm/", {
+      uid,
+      token,
+      new_password: newPassword,
+    });
+    return { success: true, detail: response.data.detail };
+  } catch (error: any) {
+    if (error.response?.data) {
+      const errorData = error.response.data;
+      if (errorData.detail) return { error: errorData.detail };
+      if (errorData.token) return { error: "El enlace es inválido o ha expirado." };
+      if (errorData.uid) return { error: "Enlace inválido." };
+      if (errorData.new_password) return { error: errorData.new_password[0] };
+    }
+    return { error: "Ocurrió un error al procesar tu solicitud." };
+  }
+}
